Set up jest and begin testing
We have to test on node and web, so multiple jest projects exist, which are segmented by file name.
 - node: `**/*.node-spec.ts`
 - web: `**/*.web-spec.ts`
 - any architecture: `**/*.spec.ts`

I also share tests between node and web often, so these are exported in a function that executes the same tests in each project. Files defining these types of tests are denoted by `**/*.shared-spec.ts` and `require`d in jest project files. I don't love this, so looking for better options